Refactor MIRACL/MrTydi 2CR into pyserini itself (#1523)
